# src/docs/swagger.yaml
definitions:
...
  CarSaleAd-Web-Api_api_dto.FileResponse:
    properties:
      description:
        type: string
      directory:
        type: string
      id:
        type: integer
      mimeType:
        type: string
      name:
        type: string
    type: object
  CarSaleAd-Web-Api_api_dto.Filter:
    properties:
      filterType:
        type: string
      from:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
...
  CarSaleAd-Web-Api_api_dto.PagedList-CarSaleAd-Web-Api_api_dto_FileResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.FileResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  CarSaleAd-Web-Api_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.Sort'
        type: array
    type: object
...
  CarSaleAd-Web-Api_api_dto.Sort:
    properties:
      col_id:
        type: string
      sort:
        type: string
    type: object
  CarSaleAd-Web-Api_api_dto.UpdateFileRequest:
    properties:
      description:
        type: string
    type: object
  CarSaleAd-Web-Api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_validations.ValidationError'
        type: array
    type: object
  CarSaleAd-Web-Api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  CarSaleAd-Web-Api_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
  api_handlers.personData:
    properties:
      first_name:
        maxLength: 10
        minLength: 4
        type: string
      last_name:
        maxLength: 20
        minLength: 6
        type: string
      mobile_number:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - first_name
    - last_name
    - mobile_number
    type: object
info:
  contact: {}
paths:
...
  /v1/files/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a File
      parameters:
      - in: formData
        name: description
        required: true
        type: string
      - description: Create a File
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: File response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a File
      tags:
      - File
  /v1/files/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a File
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a File
      tags:
      - File
    get:
      consumes:
      - application/json
      description: Get a File
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: File Response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a File
      tags:
      - File
    put:
      consumes:
      - application/json
      description: Update a File
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a File
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.UpdateFileRequest'
      responses:
        "200":
          description: File response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a File
      tags:
      - File
  /v1/files/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Files
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: File Response
          schema:
            allOf:
            - $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.PagedList-CarSaleAd-Web-Api_api_dto_FileResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Files
      tags:
      - File
...
  /v1/test/binder/body:
    post:
      consumes:
      - application/json
      description: Test BodyBinder
      parameters:
      - description: person data
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/api_handlers.personData'
      produces:
      - application/json
      responses:
        "200":
          description: Done
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: BodyBinder
      tags:
      - Test
  /v1/test/user/{id}:
    get:
      consumes:
      - application/json
      description: Test UserById
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Done
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      summary: UserById
      tags:
      - Test
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"